+++ examples/data-sources/iosxr_key_chain/data-source.tf
@@ -1,3 +1,3 @@
 data "iosxr_key_chain" "example" {
-  key_chain_name = "KEY11"
+  name = "KEY11"
 }