+++ docs/guides/changelog.md
@@ -14,6 +14,7 @@ description: |-
 - Add `iosxr_router_bgp_vrf_neighbor_address_family` resource and data source
 - Add `iosxr_router_bgp_neighbor_group_address_family` resource and data source
 - Add `iosxr_router_static` resource and data source
+- Add `iosxr_key_chain` resource and data source
 
 ## 0.1.7